Fix issue with merchants search

+++ app/src/main/java/com/bubelov/coins/ui/activity/MapActivity.java
@@ -202,7 +202,7 @@ public class MapActivity extends AbstractActivity implements OnMapReadyCallback,
 
         if (requestCode == REQUEST_FIND_MERCHANT && resultCode == RESULT_OK) {
             drawerMenu.setAmenity(null);
-            selectMerchant(data.getIntExtra(MerchantsSearchActivity.MERCHANT_ID_EXTRA, -1));
+            selectMerchant(data.getLongExtra(MerchantsSearchActivity.MERCHANT_ID_EXTRA, -1));
             map.moveCamera(CameraUpdateFactory.newLatLngZoom(selectedMerchant.getPosition(), MAP_DEFAULT_ZOOM));
         }
